Update cập nhật doanh thu tiền mặt và chuyển khoản

+++ SpringBoots/ServiceDoanhThu/src/main/java/com/example/servicedoanhthu/controller/CaController.java
@@ -51,7 +51,7 @@ public class CaController {
     }
 
     @PostMapping("/mo-ca")
-    public Ca openCa(@RequestParam BigDecimal soTienKet) {
+    public Ca openCa(@RequestParam Double soTienKet) {
         return service.openCa(soTienKet);
     }
